+++ changelog.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+
+- PlaceGrid to quickly place many shapes into a grid layout
 
 ## [0.5.0] - 2019-12-13
 ### Added